# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tool {
  id: number;
  name: string;
  slug: string;
  category: string;
  tagline: string;
  website: string;
  /** @nullable */
  logoUrl?: string | null;
  /** Developer experience score 0-10 */
  dxScore: number;
  /** Price value score 0-10 */
  priceScore: number;
  /** Frustration index 0-10 (lower is better) */
  rageIndex: number;
  /** Overall happiness score 0-10 */
  happinessScore: number;
  voteCount?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  pros?: string | null;
  /** @nullable */
  cons?: string | null;
  /** free, freemium, paid */
  pricingTier?: string;
  createdAt?: string;
}

export interface ToolInput {
  name: string;
  slug?: string;
  category: string;
  tagline: string;
  website: string;
  logoUrl?: string;
  dxScore: number;
  priceScore: number;
  rageIndex: number;
  happinessScore: number;
  description?: string;
  pros?: string;
  cons?: string;
  pricingTier: string;
}

export interface Guide {
  id: number;
  title: string;
  slug: string;
  /** guide, error-fix, tutorial */
  type: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  tags?: string | null;
  readingTime: number;
  isFeatured?: boolean;
  publishedAt: string;
  /** @nullable */
  author?: string | null;
}

export interface GuideInput {
  title: string;
  slug?: string;
  type: string;
  excerpt: string;
  content?: string;
  tags?: string;
  readingTime: number;
  isFeatured?: boolean;
  author?: string;
}

export interface HnStory {
  id: number;
  hnId?: number;
  title: string;
  url: string;
  /** @nullable */
  domain?: string | null;
  score: number;
  commentCount: number;
  /** @nullable */
  author?: string | null;
  postedAt: string;
  curatedAt?: string;
  /** @nullable */
  curatorNote?: string | null;
  /** @nullable */
  tag?: string | null;
}

export interface VoteInput {
  /** up or down */
  type: string;
}

export interface VoteResult {
  success: boolean;
  voteCount: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface StatsSummary {
  totalTools: number;
  totalGuides: number;
  totalHnStories: number;
  avgHappinessScore: number;
  topCategory?: string;
}

export type ListToolsParams = {
/**
 * Filter by category (hosting, database, payments, etc.)
 */
category?: string;
/**
 * Sort by (happiness, dx, rage, price)
 */
sort?: string;
};

export type ListGuidesParams = {
/**
 * Filter by type (guide, error-fix, tutorial)
 */
type?: string;
/**
 * Filter by tag
 */
tag?: string;
};

export type GetHnCuratedParams = {
/**
 * Number of stories to return
 */
limit?: number;
};