# lib/api-zod/src/generated/types/guide.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface Guide {
  id: number;
  title: string;
  slug: string;
  /** guide, error-fix, tutorial */
  type: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  tags?: string | null;
  readingTime: number;
  isFeatured?: boolean;
  publishedAt: string;
  /** @nullable */
  author?: string | null;
}