# lib/api-zod/src/generated/types/guideInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface GuideInput {
  title: string;
  slug?: string;
  type: string;
  excerpt: string;
  content?: string;
  tags?: string;
  readingTime: number;
  isFeatured?: boolean;
  author?: string;
}