# lib/api-zod/src/generated/types/hnStory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface HnStory {
  id: number;
  hnId?: number;
  title: string;
  url: string;
  /** @nullable */
  domain?: string | null;
  score: number;
  commentCount: number;
  /** @nullable */
  author?: string | null;
  postedAt: string;
  curatedAt?: string;
  /** @nullable */
  curatorNote?: string | null;
  /** @nullable */
  tag?: string | null;
}