# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalTools: number;
  totalGuides: number;
  totalHnStories: number;
  avgHappinessScore: number;
  topCategory?: string;
}