# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: number;
  name: string;
  slug: string;
  category: string;
  tagline: string;
  website: string;
  /** @nullable */
  logoUrl?: string | null;
  /** Developer experience score 0-10 */
  dxScore: number;
  /** Price value score 0-10 */
  priceScore: number;
  /** Frustration index 0-10 (lower is better) */
  rageIndex: number;
  /** Overall happiness score 0-10 */
  happinessScore: number;
  voteCount?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  pros?: string | null;
  /** @nullable */
  cons?: string | null;
  /** free, freemium, paid */
  pricingTier?: string;
  createdAt?: string;
}