# lib/api-zod/src/generated/types/toolInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolInput {
  name: string;
  slug?: string;
  category: string;
  tagline: string;
  website: string;
  logoUrl?: string;
  dxScore: number;
  priceScore: number;
  rageIndex: number;
  happinessScore: number;
  description?: string;
  pros?: string;
  cons?: string;
  pricingTier: string;
}