# lib/api-zod/src/generated/types/voteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Astro Dev - Daily developer resource site API
 * OpenAPI spec version: 0.1.0
 */

export interface VoteInput {
  /** up or down */
  type: string;
}